Rethink the schema and add some example routes
I've overhauled the schema, it's no longer using two separate tables as
this lead to some slow and complicated queries later. Now versions of
repositories are immutable entities and we can infer repository
information from the latest ingested version.

The example routes illustrate that data can still be exposed to users as
"repositories", and how we'd handle errors or missing data.

What's not handled yet: how to display manifests and markdown documents.
I think these may be exposed through either separate endpoints entirely
or we'll add more virtual properties to the Version model which expose
these.

The next step is to write tests for everything so far, which will appear
shortly as a new PR.