Switch to require an API key and secret pair
I've updated to make endpoints require an API key and secret pair. This
means that even if the database were compromised, there would be no need
for key rotation and we could safely restore to a previous backup.

The secrets are hashed using the `bcrypt` module. We reduce the number
of salt rounds required in order to speed up the service. Because of
this compromise, we should never use UUIDs as secrets: these could be
guessable.