Add an ingestion queue
This table, model, and set of endpoints are used to view and manage the
Ingestion Queue. This is a queue of repo URLs and tags which are waiting
to be loaded into the service.

The endpoints for listing, adding, and deleting ingestions are in place,
but there's still more to do (later):

  - Update the ingestion creation endpoint to also accept a GitHub
    webhook event. This will likely be handled in the controller, which
    can change behaviour based on the presence of an `X-GitHub-Event`
    header.

  - Add group deletion endpoints which remove any ingestions that have
    been attempted too many times, or have been running for longer than
    is acceptable.

  - Add in code which continually (or on a schedule) polls the ingestion
    queue and processes them, gathering the information required to move
    them to the versions table.