Add endpoints to retrieve manifests and markdown
I've added two new endpoints which can be used to retrieve manifest
files and markdown documents for individual versions of a repository.
This part completes #43, which explains some of the decisions made here.

I'm aware that there's a lot of repetition in the controllers. Because
the integration tests have pretty good coverage I'd rather refactor the
controllers later and split into middleware, and would feel confident
doing so. Right now it'll just slow progress and I reckon would be
premature optimisation.

What's left to do is add aliases (outlined in the original issue)