Add a new demos endpoint
I've added a new endpoint which allows you to get the normalised demo
objects for a repo/version. This is required to reduce logic in the
registry, as safely sanitizing the demos takes enough for it to be a
bit painful.

The endpoint is:

```
GET /v1/repos/:repoId/versions/:versionId/demos
```