Add an endpoint for normalised image sets
This adds a `/v1/repos/:repoId/versions/:versionId/images` endpoint
which can be used to retrieve a normalised list of images. These images
have Image Service URLs pre-baked in so that we can manage them in fewer
places and a consumer of the API doesn't have to know how the Image
Service works.

The URLs included are the full size one, and one with a width of 200px
(the width in the current registry)