Implement basic search
This is done in Node.js for now, for simplicity and ease of getting
something out to test. At some point we should maybe consider an actual
search tool or at least moving some of the processing to SQL.

There will be a second PR in Registry UI once this is merged and
released, to move the search logic out.