Use a normalised repo name to sort all repos
Sorting of repos is now a bit more considered, there are two different
ways that we sort based on the filters present.

  1. **Simple Sort**
     If no search filter is present (e.g. no filters at all, or just
     type/status filters) then we sort by a new property:
     `name_sortable` (see below)

  2. **Weighted Sort**
     If a search filter is present then we use the existing sorting
     algorithm which weights in favour of name, then keywords, then
     inferred keywords. If repos have the same search score then we
     fall back to sorting by `name_sortable`

The new `name_sortable` property is a lowercase version of the
repository name, with the team prefix removed. A team prefix is a single
lowercase letter followed by a dash at the start of the name, e.g. "o-"
or "n-". Stripping this prefix for the purposes of sorting means that
"o-ads" will appear above "ftdomdelegate", for example.