Prevent current nav items bleeding between users
We were using a single global object to represent the navigation, so
marking items as current marked them as current for every user of the
site. Oops.

This PR stops this from happening by cloning the navigation structure on
every request. I've also changed the way that we pick the top-level nav
item by testing against the `href` rather than accessing the first or
second item in the array.