Alias an origamiType of component to module
This is a first step towards supporting Origami projects having an
`origamiType` property of `component` in their manifests. Right now
we're aliasing it to `module` in repo-data, later we can switch this
behaviour but this is the simplest way to make sure that repo data and
the registry continue to work.

There's a lack of testing around our models at the moment - the
ingestion process isn't actually run in any tests which is a bit of a
worry. I think I'll remedy this separately.