Fix multiple ingestion error.
Repo data attempts to only ingest a version once by accepting new
ingestions only if the same version does not exist and is not
already in the queue. The ingestion queue is cleared on success
or a non-recoverable error.

A bug was introduced in `Version.createFromIngestion` where the
new version was not returned. This meant a new version was created
but the queue wasn't cleared, due to an error. The queue would then
create a new version repeatedly. The error thrown should have been
caught, and the queue cleared, but as the error wasn't explicitly marked
as non-recoverable the queue was not cleared.
https://github.com/Financial-Times/origami-repo-data/pull/305/files#diff-a786e1980e01b7d9ccf2ed922047ab447189cdb2e89776b6b6b2bf98246b375eR587

Changes:
- Return the version.
- Clear the queue on error _unless_ the error is explicitly marked as
recoverable.
- Report caught ingestion errors to Sentry.

Ideally we'd improve test coverage and discuss introducing types to
catch this kind of issue, but let's see if we can decommission repo
data instead.