Drop the "module" origami type for spec v2 components.
- Continue to normalise "component" to "module" for spec v1 components, which may
use either. "component" was added as an alias later, and we want any tooling to
be able to continue to work with v1 component results.
- Swap "module" for "component" as a type for spec v2 tests and seed data. Tests
for the filter query param show fewer results correctly, as the query is for
"module" which no longer returns spec v2 components.
- Create `Version` `type_is_component` virtual property so repo data may treat
v1 and v2 components the same by querying this property.

Conversation:
https://github.com/Financial-Times/origami-website/pull/273#discussion_r617396987